Rename service infra-proxy to infra-proxy-service
