--- components/infra-proxy/package.meta
@@ -1,4 +1,0 @@
-{
-  "name": "chef/infra-proxy",
-  "data_service": true
-}
